% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bear1d.R
\name{calculate_bear1d}
\alias{calculate_bear1d}
\title{Calculate Bear 1D}
\usage{
calculate_bear1d(
  n_values,
  rs_values,
  foc_values,
  log_koc_values,
  hl_values,
  C0 = 1,
  D_values,
  v_values,
  t_values,
  x_values
)
}
\arguments{
\item{n_values}{Porosity (-)}

\item{rs_values}{density solids (g/g)}

\item{foc_values}{fraction of organic matter}

\item{log_koc_values}{min, max Koc for relevant compounds Nguyen et al 2020}

\item{hl_values}{half life (T)}

\item{C0}{initial concentration of the solute (M/L3)}

\item{D_values}{coefficient of longitudinal dispersion (L2/T)}

\item{v_values}{average linear ground water velocity (L/T)}

\item{t_values}{time (T)}

\item{x_values}{flow path distance (L)}
}
\value{
Bear 1D results
}
\description{
Calculate Bear 1D
}
\examples{
bear1d_vienna <- calculate_bear1d(n_values = 0.15,
rs_values = 2.7,
foc_values = 0.002,
log_koc_values = c(1,2.9),
hl_values = c(500,2000),
C0 = 1,
D_values = 10,
v_values = 21.3,
t_values = seq(0, 350, by=1),
x_values = 141
)
}
