% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Cx.R
\name{calculate_Cx}
\alias{calculate_Cx}
\title{Calculate Cx}
\usage{
calculate_Cx(x, t, v, D, retardation, C0, k)
}
\arguments{
\item{x}{x}

\item{t}{t}

\item{v}{v}

\item{D}{D}

\item{retardation}{retardation}

\item{C0}{initial concentration}

\item{k}{k}
}
\value{
????
}
\description{
Calculate Cx
}
