% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Cx.R
\name{calculate_e}
\alias{calculate_e}
\title{Helper function: calculate e}
\usage{
calculate_e(k, D, retardation, v, t)
}
\arguments{
\item{k}{k}

\item{D}{D}

\item{retardation}{retardation}

\item{v}{v}

\item{t}{t}
}
\value{
???
}
\description{
Helper function: calculate e
}
\keyword{internal}
