% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Cx.R
\name{calculate_f}
\alias{calculate_f}
\title{Helper function: calculate f}
\usage{
calculate_f(D, retardation, t)
}
\arguments{
\item{D}{D}

\item{retardation}{retardation}

\item{t}{t}
}
\value{
???
}
\description{
Helper function: calculate f
}
\keyword{internal}
