% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Cx.R
\name{calculate_g}
\alias{calculate_g}
\title{Helper function: calculate g}
\usage{
calculate_g(x, v, D)
}
\arguments{
\item{x}{x}

\item{v}{v}

\item{D}{D}
}
\value{
???
}
\description{
Helper function: calculate g
}
\keyword{internal}
