% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combined.R
\name{plot_combined}
\alias{plot_combined}
\title{Plot Combined}
\usage{
plot_combined(bear1d_list)
}
\arguments{
\item{bear1d_list}{Bear1D list as retrieved by \code{\link{calculate_bear1d}}}
}
\value{
combined plot
}
\description{
Plot Combined
}
