% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(bear1d_list)
}
\arguments{
\item{bear1d_list}{Bear1D list as retrieved by \code{\link{calculate_bear1d}}}
}
\value{
heatmap plot
}
\description{
Plot Heatmap
}
